requirements.txt

streamlit
openai
sentence-transformers
faiss-cpu
requests
python-dotenv
